+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-EasyEdit is an easy markdown editor that allows you to write Markdown (MD) and preview it in real-time. You can save, load .md files and export to HTML & PDF. All notable changes to this project will be documented in this file.
+EasyEdit is an easy markdown editor that allows you to write Markdown (MD) and preview it in real-time. You can save, load .md files and export to HTML,TXT & PDF. All notable changes to this project will be documented in this file.
 
 ## Latest version
-- 1.1.0 - Additional graph plaintext and flowchart
+- 1.1.1 - Added export Text files and improved mermaid code
 
 ## Latest built version
 - 1.1.0 - Linux (AppImage)
@@ -26,3 +26,4 @@ EasyEdit is an easy markdown editor that allows you to write Markdown (MD) and p
 - 1.0.8 - Implemented HTML code format and Export Markdown to HTML support
 - 1.0.9 - To improve development of saveAsPDF function the code has been moved into separate file saveAsPDF.tsx
 - 1.1.0 - Additional graph plaintext and flowchart
+- 1.1.1 - Added export Text files and improved mermaid code